requirements.txt
Flask==3.0.0
DrissionPage
Pillow
requests